third_party/WebKit/LayoutTests/svg/parser/whitespace-length-expected.txt

This is a testharness.js-based test.
FAIL Test <length> trailing garbage, value: -47 assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: .1 assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: 0.35 assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: 1e-10
FAIL Test <length> trailing garbage, value: +32 assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: +17E-1 assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: 17e+2 assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: -47em assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: .1em assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: 0.35em assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: 1e-10em
FAIL Test <length> trailing garbage, value: +32em assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: +17E-1em assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: 17e+2em assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: -47ex assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: .1ex assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: 0.35ex assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: 1e-10ex
FAIL Test <length> trailing garbage, value: +32ex assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: +17E-1ex assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: 17e+2ex assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: -47px assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: .1px assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: 0.35px assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: 1e-10px
FAIL Test <length> trailing garbage, value: +32px assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: +17E-1px assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: 17e+2px assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: -47in assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: .1in assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: 0.35in assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: 1e-10in
FAIL Test <length> trailing garbage, value: +32in assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: +17E-1in assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: 17e+2in assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: -47cm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: .1cm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: 0.35cm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: 1e-10cm
FAIL Test <length> trailing garbage, value: +32cm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: +17E-1cm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: 17e+2cm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: -47mm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: .1mm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: 0.35mm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: 1e-10mm
FAIL Test <length> trailing garbage, value: +32mm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: +17E-1mm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: 17e+2mm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: -47pt assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: .1pt assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: 0.35pt assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: 1e-10pt
FAIL Test <length> trailing garbage, value: +32pt assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: +17E-1pt assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: 17e+2pt assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: -47pc assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: .1pc assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: 0.35pc assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: 1e-10pc
FAIL Test <length> trailing garbage, value: +32pc assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: +17E-1pc assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: 17e+2pc assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: -47% assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: .1% assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: 0.35% assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: 1e-10%
FAIL Test <length> trailing garbage, value: +32% assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: +17E-1% assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: 17e+2% assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:  -47 assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:  .1 assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:  0.35 assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:  1e-10
FAIL Test <length> trailing garbage, value:  +32 assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:  +17E-1 assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:  17e+2 assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:  -47em assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:  .1em assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:  0.35em assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:  1e-10em
FAIL Test <length> trailing garbage, value:  +32em assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:  +17E-1em assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:  17e+2em assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:  -47ex assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:  .1ex assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:  0.35ex assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:  1e-10ex
FAIL Test <length> trailing garbage, value:  +32ex assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:  +17E-1ex assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:  17e+2ex assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:  -47px assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:  .1px assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:  0.35px assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:  1e-10px
FAIL Test <length> trailing garbage, value:  +32px assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:  +17E-1px assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:  17e+2px assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:  -47in assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:  .1in assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:  0.35in assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:  1e-10in
FAIL Test <length> trailing garbage, value:  +32in assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:  +17E-1in assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:  17e+2in assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:  -47cm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:  .1cm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:  0.35cm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:  1e-10cm
FAIL Test <length> trailing garbage, value:  +32cm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:  +17E-1cm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:  17e+2cm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:  -47mm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:  .1mm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:  0.35mm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:  1e-10mm
FAIL Test <length> trailing garbage, value:  +32mm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:  +17E-1mm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:  17e+2mm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:  -47pt assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:  .1pt assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:  0.35pt assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:  1e-10pt
FAIL Test <length> trailing garbage, value:  +32pt assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:  +17E-1pt assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:  17e+2pt assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:  -47pc assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:  .1pc assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:  0.35pc assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:  1e-10pc
FAIL Test <length> trailing garbage, value:  +32pc assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:  +17E-1pc assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:  17e+2pc assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:  -47% assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:  .1% assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:  0.35% assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:  1e-10%
FAIL Test <length> trailing garbage, value:  +32% assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:  +17E-1% assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:  17e+2% assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:    -47 assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:    .1 assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:    0.35 assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:    1e-10
FAIL Test <length> trailing garbage, value:    +32 assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:    +17E-1 assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:    17e+2 assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:    -47em assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:    .1em assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:    0.35em assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:    1e-10em
FAIL Test <length> trailing garbage, value:    +32em assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:    +17E-1em assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:    17e+2em assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:    -47ex assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:    .1ex assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:    0.35ex assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:    1e-10ex
FAIL Test <length> trailing garbage, value:    +32ex assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:    +17E-1ex assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:    17e+2ex assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:    -47px assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:    .1px assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:    0.35px assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:    1e-10px
FAIL Test <length> trailing garbage, value:    +32px assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:    +17E-1px assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:    17e+2px assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:    -47in assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:    .1in assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:    0.35in assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:    1e-10in
FAIL Test <length> trailing garbage, value:    +32in assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:    +17E-1in assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:    17e+2in assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:    -47cm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:    .1cm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:    0.35cm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:    1e-10cm
FAIL Test <length> trailing garbage, value:    +32cm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:    +17E-1cm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:    17e+2cm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:    -47mm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:    .1mm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:    0.35mm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:    1e-10mm
FAIL Test <length> trailing garbage, value:    +32mm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:    +17E-1mm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:    17e+2mm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:    -47pt assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:    .1pt assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:    0.35pt assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:    1e-10pt
FAIL Test <length> trailing garbage, value:    +32pt assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:    +17E-1pt assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:    17e+2pt assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:    -47pc assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:    .1pc assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:    0.35pc assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:    1e-10pc
FAIL Test <length> trailing garbage, value:    +32pc assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:    +17E-1pc assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:    17e+2pc assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:    -47% assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:    .1% assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:    0.35% assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:    1e-10%
FAIL Test <length> trailing garbage, value:    +32% assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:    +17E-1% assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:    17e+2% assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \r\n\t -47 assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \r\n\t .1 assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \r\n\t 0.35 assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \r\n\t 1e-10
FAIL Test <length> trailing garbage, value: \r\n\t +32 assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \r\n\t +17E-1 assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \r\n\t 17e+2 assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \r\n\t -47em assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \r\n\t .1em assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \r\n\t 0.35em assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \r\n\t 1e-10em
FAIL Test <length> trailing garbage, value: \r\n\t +32em assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \r\n\t +17E-1em assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \r\n\t 17e+2em assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \r\n\t -47ex assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \r\n\t .1ex assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \r\n\t 0.35ex assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \r\n\t 1e-10ex
FAIL Test <length> trailing garbage, value: \r\n\t +32ex assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \r\n\t +17E-1ex assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \r\n\t 17e+2ex assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \r\n\t -47px assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \r\n\t .1px assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \r\n\t 0.35px assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \r\n\t 1e-10px
FAIL Test <length> trailing garbage, value: \r\n\t +32px assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \r\n\t +17E-1px assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \r\n\t 17e+2px assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \r\n\t -47in assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \r\n\t .1in assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \r\n\t 0.35in assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \r\n\t 1e-10in
FAIL Test <length> trailing garbage, value: \r\n\t +32in assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \r\n\t +17E-1in assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \r\n\t 17e+2in assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \r\n\t -47cm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \r\n\t .1cm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \r\n\t 0.35cm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \r\n\t 1e-10cm
FAIL Test <length> trailing garbage, value: \r\n\t +32cm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \r\n\t +17E-1cm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \r\n\t 17e+2cm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \r\n\t -47mm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \r\n\t .1mm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \r\n\t 0.35mm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \r\n\t 1e-10mm
FAIL Test <length> trailing garbage, value: \r\n\t +32mm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \r\n\t +17E-1mm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \r\n\t 17e+2mm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \r\n\t -47pt assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \r\n\t .1pt assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \r\n\t 0.35pt assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \r\n\t 1e-10pt
FAIL Test <length> trailing garbage, value: \r\n\t +32pt assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \r\n\t +17E-1pt assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \r\n\t 17e+2pt assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \r\n\t -47pc assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \r\n\t .1pc assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \r\n\t 0.35pc assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \r\n\t 1e-10pc
FAIL Test <length> trailing garbage, value: \r\n\t +32pc assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \r\n\t +17E-1pc assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \r\n\t 17e+2pc assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \r\n\t -47% assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \r\n\t .1% assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \r\n\t 0.35% assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \r\n\t 1e-10%
FAIL Test <length> trailing garbage, value: \r\n\t +32% assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \r\n\t +17E-1% assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \r\n\t 17e+2% assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \f-47 assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \f.1 assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \f0.35 assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \f1e-10
FAIL Test <length> trailing garbage, value: \f+32 assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \f+17E-1 assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \f17e+2 assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \f-47em assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \f.1em assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \f0.35em assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \f1e-10em
FAIL Test <length> trailing garbage, value: \f+32em assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \f+17E-1em assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \f17e+2em assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \f-47ex assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \f.1ex assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \f0.35ex assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \f1e-10ex
FAIL Test <length> trailing garbage, value: \f+32ex assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \f+17E-1ex assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \f17e+2ex assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \f-47px assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \f.1px assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \f0.35px assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \f1e-10px
FAIL Test <length> trailing garbage, value: \f+32px assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \f+17E-1px assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \f17e+2px assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \f-47in assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \f.1in assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \f0.35in assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \f1e-10in
FAIL Test <length> trailing garbage, value: \f+32in assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \f+17E-1in assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \f17e+2in assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \f-47cm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \f.1cm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \f0.35cm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \f1e-10cm
FAIL Test <length> trailing garbage, value: \f+32cm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \f+17E-1cm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \f17e+2cm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \f-47mm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \f.1mm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \f0.35mm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \f1e-10mm
FAIL Test <length> trailing garbage, value: \f+32mm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \f+17E-1mm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \f17e+2mm assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \f-47pt assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \f.1pt assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \f0.35pt assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \f1e-10pt
FAIL Test <length> trailing garbage, value: \f+32pt assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \f+17E-1pt assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \f17e+2pt assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \f-47pc assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \f.1pc assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \f0.35pc assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \f1e-10pc
FAIL Test <length> trailing garbage, value: \f+32pc assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \f+17E-1pc assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \f17e+2pc assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \f-47% assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \f.1% assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \f0.35% assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \f1e-10%
FAIL Test <length> trailing garbage, value: \f+32% assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \f+17E-1% assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \f17e+2% assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
Harness: the test ran to completion.
